Simplify template and provide IR example
Simplify the directory tree so that ACS results get copied into a top
level folder and is easier to find. Remove the remaining /acs/*
directories like `fwts` and `sct` to avoid confusion about where results
need to go. Instead create an empty top level /manual-results/ tree
which is where manual test results should be placed, but don't
pre-populate the directories because in most cases this is not needed
and instructions are covered in the README.

Also update the README to reflect the new guidance and to list
directories with a trailing '/' so that they are recognized as
directories and not files.

Signed-off-by: Grant Likely <grant.likely@arm.com>